requirements.txt
# Core dependencies
pydantic>=2.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
PyMuPDF>=1.22.0
python-docx>=0.8.11
openai>=1.0.0
anthropic>=0.25.0
pandas>=2.0.0
pyarrow>=12.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
pydantic-settings>=2.0.0

psutil>=5.9.0
